Updated versions of Sporth and Soundpipe

+++ AudioKit/Common/Internals/Soundpipe/modules/fof.c
@@ -53,7 +53,7 @@ static int newpulse(sp_data *sp,
     ovp->timrem = p->dur * sp->sr;
 
     if ((oct = p->oct) > 0.0) {
-        int32_t ioct = (int32_t)oct, bitpat = ~(-1L << ioct);
+        int32_t ioct = (int32_t)oct, bitpat = (int) ~(-1L << ioct);
         if (bitpat & ++p->fofcount) return(0);
         if ((bitpat += 1) & p->fofcount) octamp *= (1.0 + ioct - oct);
     }
